Fetch Customer details using ajax/jquery Updated headers - imports in js

+++ database/migrations/2020_04_12_141607_create_products.php
@@ -16,7 +16,7 @@ class CreateProducts extends Migration
         Schema::create('products', function (Blueprint $table) {
             $table->bigIncrements('id');
             $table->string('name');
-            $table->float('price', 5, 2);
+            $table->decimal('price', 5, 2);
             $table->text('details')->nullable();
             $table->unsignedBigInteger('product_type_id')->unsigned();
             $table->timestamps();
